Move date and time functions into helper.js.
